+++ Category.UIKIt/UIImage+StackBlur.m
@@ -60,7 +60,7 @@ inline static void zeroClearInt(int* p, size_t count) { memset(p, 0, sizeof(int)
     //	return [other applyBlendFilter:filterOverlay  other:self context:nil];
 	// First get the image into your data buffer
     CGImageRef inImage = self.CGImage;
-    int nbPerCompt = CGImageGetBitsPerPixel(inImage);
+    size_t nbPerCompt = CGImageGetBitsPerPixel(inImage);
     if(nbPerCompt != 32){
         UIImage *tmpImage = [self normalize];
         inImage = tmpImage.CGImage;
@@ -83,8 +83,8 @@ inline static void zeroClearInt(int* p, size_t count) { memset(p, 0, sizeof(int)
 											 );
 	
     // Apply stack blur
-    const int imageWidth  = CGImageGetWidth(inImage);
-	const int imageHeight = CGImageGetHeight(inImage);
+    const size_t imageWidth  = CGImageGetWidth(inImage);
+	const size_t imageHeight = CGImageGetHeight(inImage);
     [self.class applyStackBlurToBuffer:m_PixelBuf
                                  width:imageWidth
                                 height:imageHeight
